Partial splitting/joining of lines.
You know that problem when you're trying to fix word wrapping, and you
keep manually moving words at the end of the line up or down? It's super
annoying. These mappings make that easier by letting you move a single
word (or a group of words) at a time, either up or down.

ALT+j: Split Below - Make the last word of this line the first word of
the next line.
ALT+J: Join Below - Make the first word of the next line the last word
of this line.
ALT+k: Split Above - Make the first word of this line the last word of
the previous line.
ALT+K: Join Above - Make the last word of the previous line the first
word of this line.

This works with a count prefix, e.g. 3+ALT+j will make the
last 3 words of this line the first 3 words of the next line.

Note that this does not currently support the . (repeat) operation. It
could, but then there'd be a repeat.vim dependency.